requirements.txt
playwright>=1.43.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
